refactor: change directory and add @Deprecated
